# kask/docs/generated/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "hKask API",
    "description": "A Minimal Viable Container for Replicants — HTTP API.\n\nhKask is an agent runtime grounded in 12 architectural principles\n(P0–P12) expressed through four composable patterns: Skills Model,\nRegulation Feedback Loop, Agentic AI Mediation, and Agent Creation with\nSovereign Memory. This API exposes all capabilities equally across\nCLI, API, and MCP surfaces under P3 (Equal Surface Exposure).\n\nAll endpoints carry OCAP DelegationToken authentication (P4).\nData access is governed by user sovereignty and affirmative\nconsent (P1, P2).",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "version": "0.30.0"
  },
  "servers": [
    {
      "url": "/api",
      "description": "hKask API server"
    }
  ],
  "paths": {
...
    "/api/regulation/subscribe": {
      "get": {
        "tags": [
          "regulation"
        ],
        "summary": "Subscribe to Regulation events as an SSE stream.",
        "description": "The endpoint upgrades the HTTP response to a long-lived SSE connection.\nEvents matching the requested span namespaces are forwarded in real time.\nLag notifications are emitted when the client falls behind.",
        "operationId": "reg_subscribe",
        "parameters": [
          {
            "name": "spans",
            "in": "query",
            "description": "Span namespaces to subscribe to (e.g., [\"reg.tool\", \"reg.inference\"])",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "SSE event stream",
            "content": {
              "text/event-stream": {}
            }
          },
          "400": {
            "description": "Invalid request"
          }
        }
      }
    },
    "/api/regulation/variety": {
      "get": {
        "tags": [
          "regulation"
        ],
        "summary": "Regulation variety endpoint",
        "operationId": "reg_variety",
        "responses": {
          "200": {
            "description": "Regulation variety counters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegVarietyResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "RegVarietyResponse": {
        "type": "object",
        "description": "Regulation variety response — per-domain variety counters for the Regulation System (Pattern B).\n\n`domains` lists all canonical Regulation span namespaces registered (e.g., reg.tool,\nreg.inference, reg.memory). `total_deficit` is the aggregate variety gap across\nall domains.",
        "required": [
          "domains",
          "total_deficit",
          "counters"
        ],
        "properties": {
          "counters": {
            "type": "object",
            "description": "Per-domain variety counters keyed by namespace",
            "additionalProperties": {
              "$ref": "#/components/schemas/VarietyCounterResponse"
            },
            "propertyNames": {
              "type": "string"
            }
          },
          "domains": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "All registered Regulation span namespace domains"
          },
          "total_deficit": {
            "type": "integer",
            "format": "int64",
            "description": "Aggregate variety deficit across all domains",
            "minimum": 0
          }
        }
      },
...
      "VarietyCounterResponse": {
        "type": "object",
        "description": "Regulation variety counter for a single domain.\n\n`variety` is the tracked behavioral diversity for this domain.\n`entropy` is the Shannon entropy of the domain's observation distribution\n(0.0 when not yet computed).",
        "required": [
          "variety",
          "total",
          "entropy"
        ],
        "properties": {
          "entropy": {
            "type": "number",
            "format": "double",
            "description": "Shannon entropy (0.0 = not computed yet)"
          },
          "total": {
            "type": "integer",
            "format": "int64",
            "description": "Total observations in this domain",
            "minimum": 0
          },
          "variety": {
            "type": "integer",
            "format": "int64",
            "description": "Tracked behavioral diversity count",
            "minimum": 0
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer_token": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "DelegationToken",
        "description": "DelegationToken — an OCAP capability token carrying the authenticated WebID and scoped permissions (P4).\n\nObtain via the REPL onboarding flow (`kask secret`) or agent registration (A2A)."
      }
    }
  },
  "security": [
    {
      "bearer_token": []
    }
  ],
...
}